Adjust the output range and fix carrying To make the output range of PWM match the original analogWrite() from Arduino, it's necessary to adjust the code. The original duty cycle of PWM output of ATmega328P is calculated as (n+1)/2^16. In analogWrite() of Arduino, the code of 0x0000 represents outputting the logic low.
This patch fixes the output range so that it matches the behavior of the original analogWrite(). Additionally, it also addresses a distortion problem that occurred when carrying from LowHalf to HighHalf. The original code caused unstable output due to the flip-flop not being cleared when carrying on Channel1.

It's important to note that at the moment, it's not possible to set the duty cycle below 256 on Channel1. This should be fixed in the future.